// chaincode/contract/src/model/mReqMintNFTTicket.ts
/**
 * Ticket System components definiton
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ENFTTicket } from './eNFTTicket';
import { TRequestHeader } from './tRequestHeader';

/**
 * Mints a NFTTicket based on provided parameters
 */
export interface MReqMintNFTTicket { 
    header?: TRequestHeader;
    document?: ENFTTicket;
}