// chaincode/contract/src/model/mResStatus.ts
/**
 * Ticket System components definiton
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TResponseHeader } from './tResponseHeader';
import { TStatus } from './tStatus';

/**
 * General status
 */
export interface MResStatus { 
    header?: TResponseHeader;
    status?: TStatus;
}