// chaincode/contract/src/model/mResEvent.ts
/**
 * Ticket System components definiton
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EEvent } from './eEvent';
import { TResponseHeader } from './tResponseHeader';

/**
 * Event's data
 */
export interface MResEvent { 
    header?: TResponseHeader;
    document?: EEvent;
}